+++ src/main/java/com/omgm/user/controller/OneController.java
@@ -45,54 +45,11 @@ public class OneController {
         return mav;
     }
 
-    // 이용후기 글쓰기 페이지로 이동
-    @RequestMapping(value="/reviewWrite.do")
-    public ModelAndView reviewWrite(BoardVO vo) {
-        ModelAndView mav = new ModelAndView();
-        mav.setViewName("/review/reviewWrite");
-        return mav;
-    }
-
-    //이용후기 리스트 페이지 이동
-    @RequestMapping(value="/reviewListBoard.do")
-    public ModelAndView reviewListBoard(BoardVO vo) {
-        ModelAndView mav = new ModelAndView();
-
-        mav.addObject("board",boardService.getBoard(vo));
-        mav.setViewName("/review/reviewList_board");
-
-        mav.setViewName("/review/reviewListBoard");
-
-        return mav;
-    }
-
-    // 이용후기 본문 페이지 이동
-    @RequestMapping(value="/reviewContent.do")
-    public ModelAndView reviewContent(BoardVO vo) {
-        ModelAndView mav = new ModelAndView();
-        mav.setViewName("/review/reviewContent");
-        return mav;
-    }
-
     // 픽업서비스 소개 페이지 이동
     @RequestMapping(value="/pickupService_info.do")
     public ModelAndView pickupService_info(BoardVO vo) {
         ModelAndView mav = new ModelAndView();
         mav.setViewName("/pickupService_info/pickupService_info");
-        return mav;
-    }
-
-    // 이용후기 댓글 주고받기
-    @ResponseBody
-    @RequestMapping(value = "/reviewContentReply.do")
-    public ModelAndView map(@RequestBody ReviewVO vo) {
-        ModelAndView mav = new ModelAndView();
-
-        // DB INSERT Reply 작업
-        System.out.println(vo.getWriter());
-        System.out.println(vo.getPass());
-        System.out.println(vo.getContent());
-        mav.setViewName("/review/reviewContent");
         return mav;
     }
