+++ src/main/webapp/index.jsp
@@ -5,22 +5,19 @@
   </head>
   <body>
 
-  <a href="main.do"><h3>프로젝트 메인화면</h3></a>
-  <a href="adminMain.mdo"><h3>관리자 메인 화면 이동</h3></a>
-  <hr>
-  <a href="board.do"><h3>반응형 게시판 이동</h3></a>
-  <a href="kinderGarden.do"><h3>유치원 소개 페이지</h3></a>
-  <hr>
-  <h1>Review 페이지</h1>
-  <a href="reviewWrite.do"><h3>review_writer 페이지 이동</h3></a>
-  <a href="reviewListBoard.do"><h3>review_list 페이지 이동</h3></a>
-  <a href="reviewContent.do">review_content 페이지 이동</a>
-  <hr>
-  <h1>개별 작업 페이지</h1>
-  <a href="messageService.do"><h3>문자 서비스 소개 페이지</h3></a>
-  <a href="pickupService_info.do"><h3>pickup-intro 페이지로 이동</h3></a>
-  <a href="kinderGarden.do"><h3>유치원 소개 페이지</h3></a>
-  <a href="reservation.do"><h3>호텔 예약 페이지</h3></a>
-  
+    <a href="main.do"><h3>프로젝트 메인화면</h3></a>
+    <a href="adminMain.mdo"><h3>관리자 메인 화면 이동</h3></a>
+    <hr>
+    <h1>Review 페이지</h1>
+    <a href="reviewWrite.do"><h3>review_writer 페이지 이동</h3></a>
+    <a href="reviewListBoard.do"><h3>review_list 페이지 이동</h3></a>
+    <a href="reviewContent.do">review_content 페이지 이동</a>
+    <hr>
+    <h1>개별 작업 페이지</h1>
+    <a href="messageService.do"><h3>문자 서비스 소개 페이지</h3></a>
+    <a href="pickupService_info.do"><h3>pickup-intro 페이지로 이동</h3></a>
+    <a href="kinderGarden.do"><h3>유치원 소개 페이지</h3></a>
+    <a href="reservation.do"><h3>호텔 예약 페이지</h3></a>
+
   </body>
 </html>
